Move NSwagStudio options to its own namespace
