+++ src/ApiClientCodeGen.Tests/Generators/NSwagStudio/NSwagStudioFileHelperTests.cs
@@ -5,6 +5,7 @@ using Moq;
 using System.IO;
 using System.Threading.Tasks;
 using ChristianHelle.DeveloperTools.CodeGenerators.ApiClient.Options.NSwag;
+using ChristianHelle.DeveloperTools.CodeGenerators.ApiClient.Options.NSwagStudio;
 
 namespace ChristianHelle.DeveloperTools.CodeGenerators.ApiClient.Tests.Generators.NSwagStudio
 {
