+++ src/ApiClientCodeGen.VSIX/Generators/NSwagStudio/NSwagStudioFileHelper.cs
@@ -2,6 +2,7 @@
 using ChristianHelle.DeveloperTools.CodeGenerators.ApiClient.Extensions;
 using ChristianHelle.DeveloperTools.CodeGenerators.ApiClient.Options;
 using ChristianHelle.DeveloperTools.CodeGenerators.ApiClient.Options.NSwag;
+using ChristianHelle.DeveloperTools.CodeGenerators.ApiClient.Options.NSwagStudio;
 using NJsonSchema.CodeGeneration.CSharp;
 using NSwag;
