+++ src/ApiClientCodeGen.VSIX/Options/NSwagStudio/INSwagStudioOptions.cs
@@ -1,4 +1,6 @@
-﻿namespace ChristianHelle.DeveloperTools.CodeGenerators.ApiClient.Options.NSwag
+﻿using ChristianHelle.DeveloperTools.CodeGenerators.ApiClient.Options.NSwag;
+
+namespace ChristianHelle.DeveloperTools.CodeGenerators.ApiClient.Options.NSwagStudio
 {
     public interface INSwagStudioOptions : INSwagOptions
     {
