+++ src/ApiClientCodeGen.VSIX/Options/NSwagStudio/NSwagStudioOptionsPage.cs
@@ -3,7 +3,7 @@ using System.Diagnostics.CodeAnalysis;
 using Microsoft.VisualStudio.Shell;
 using NJsonSchema.CodeGeneration.CSharp;
 
-namespace ChristianHelle.DeveloperTools.CodeGenerators.ApiClient.Options.NSwag
+namespace ChristianHelle.DeveloperTools.CodeGenerators.ApiClient.Options.NSwagStudio
 {
     [ExcludeFromCodeCoverage]
     public class NSwagStudioOptionsPage : DialogPage, INSwagStudioOptions
